Moved all map_to() implementations into Utility.hpp
