+++ Applications/Spire/Source/BookView/QuotePanelIndicatorWidget.cpp
@@ -1,15 +1,12 @@
 #include "Spire/BookView/QuotePanelIndicatorWidget.hpp"
 #include <QPainter>
+#include "Spire/Spire/Utility.hpp"
 
 using namespace Spire;
 
 namespace {
   constexpr auto FADE_TIME_MS = 400;
   constexpr auto STARTING_ALPHA = 100;
-
-  double map_to(double value, double a1, double a2, double b1, double b2) {
-    return ((value - a1) * ((b2 - b1) / (a2 - a1))) + b1;
-  }
 }
 
 QuotePanelIndicatorWidget::QuotePanelIndicatorWidget(QWidget* parent)
@@ -44,7 +41,8 @@ void QuotePanelIndicatorWidget::on_animation_timer() {
     m_color.setAlpha(255);
   } else {
     m_color.setAlpha(static_cast<int>(map_to(
-      static_cast<double>(elapsed.count()), 0, FADE_TIME_MS, STARTING_ALPHA,
+      static_cast<double>(elapsed.count()), static_cast<double>(0),
+      static_cast<double>(FADE_TIME_MS), STARTING_ALPHA,
       255)));
     m_animation_timer.start();
   }
